fix(rtc): 修复会议票据验证逻辑
- 将票据中的用户 ID 和成员 ID 从 Integer 转换为 Long
- 修复了由于类型不匹配导致的验证失败问题

+++ ruoyi-modules/ruoyi-rtc/src/main/java/com/ruoyi/rtc/pojo/handle/JoinSignal.java
@@ -85,8 +85,8 @@ public class JoinSignal extends SignalBase {
         if (StringUtils.isNotEmpty(ticket)) {
             Claims claims = JwtUtils.parseToken(ticket);
             String meetingIdInTicket = (String) claims.get(MEETING_TICKET_ITEM_MEETING_ID);
-            Long userIdInTicket = (Long) claims.get(MEETING_TICKET_ITEM_MEETING_OWNER_USER_ID);
-            Long memberIdInTicket = (Long) claims.get(MEETING_TICKET_ITEM_MEETING_MEMBER_USER_ID);
+            Long userIdInTicket = ((Integer) claims.get(MEETING_TICKET_ITEM_MEETING_OWNER_USER_ID)).longValue();
+            Long memberIdInTicket = ((Integer) claims.get(MEETING_TICKET_ITEM_MEETING_MEMBER_USER_ID)).longValue();
 
             // 票据校验正确 JOIN_RESOLVE
             if (meetingId.equals(meetingIdInTicket) && currentUserId.equals(memberIdInTicket) && meeting.getUserId().equals(userIdInTicket)) {
